ML/requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Web Framework
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0

# Visualization (optional but recommended)
matplotlib==3.7.2
seaborn==0.12.2

# Advanced ML Libraries (optional - uncomment if needed)
# xgboost==2.0.3
# lightgbm==4.1.0
# optuna==3.4.0

# Model Interpretability (optional)
# shap==0.43.0

# Testing (optional)
# pytest==7.4.3